- Added new transformers for embedding and dimensionality reduction. - Introduced `.pre-commit-config.yaml` to enforce code formatting and linting using Ruff. - Updated `mkdocs.yml` to include new sections for Meta-Learning Estimators and DSPyMator in the documentation. - Added `umap-learn` and `pytest-cov` to `pyproject.toml` for enhanced functionality and testing. - Improved installation instructions in `README.md` and `installation-quick-start.md`. - Added new API reference documentation for DSPyMator and KerasCortex. - Refactored feature transformers and model estimators for better organization and clarity. - Enhanced CI workflows for testing and documentation deployment.
