+++ .github/workflows/ci.yml
@@ -1,37 +1,43 @@
-name: ci 
+name: ci
 on:
   push:
-    branches:
-      - master 
-      - main
-permissions:
-  contents: write
+    branches: [main, master]
+  pull_request:
+
 jobs:
-  deploy:
+  test:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.11", "3.12"]
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install uv
+        run: curl -LsSf https://astral.sh/uv/install.sh | sh
+      - name: Install dependencies
+        run: uv pip install -e .[all] pytest pytest-cov --system
+      - name: Run tests with coverage
+        run: uv run pytest --cov=centimators --cov-report=term --cov-fail-under=80
+
+  lint:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-      - name: Configure Git Credentials
-        run: |
-          git config user.name github-actions[bot]
-          git config user.email 41898282+github-actions[bot]@users.noreply.github.com
-      - uses: actions/setup-python@v5
-        with:
-          python-version: 3.x
-      - run: echo "cache_id=$(date --utc '+%V')" >> $GITHUB_ENV 
-      - uses: actions/cache@v4
-        with:
-          key: mkdocs-material-${{ env.cache_id }}
-          path: .cache 
-          restore-keys: |
-            mkdocs-material-
-      - name: Install uv
-        run: |
-          curl -LsSf https://astral.sh/uv/install.sh | sh
-          echo "$HOME/.cargo/bin" >> $GITHUB_PATH
-      - name: Install dependencies
-        run: |
-          uv pip install mkdocs-material mkdocstrings[python] jupyter mkdocs-jupyter polars markdown-exec[ansi] --system
-          # Install the package itself
-          uv pip install -e . --system
-      - run: mkdocs gh-deploy --force
+      - uses: astral-sh/setup-uv@v4
+      - name: Check formatting
+        run: uv run ruff format --check .
+      - name: Lint
+        run: uv run ruff check --exit-non-zero-on-fix .
+
+  type:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: astral-sh/setup-uv@v4
+      - name: Type check
+        run: uv run ty check src/
+
+
