+++ docs/api-reference/dspymator.md
@@ -1,3 +1,3 @@
-::: centimators.feature_transformers
+::: centimators.model_estimators.dspymator
     options:
       show_symbol_type_toc: true
