+++ docs/api-reference/losses.md
@@ -1,8 +1,3 @@
-# Losses
-
-::: centimators.losses.SpearmanCorrelation
-    options:
-      show_symbol_type_toc: true
-::: centimators.losses.CombinedLoss
+::: centimators.losses
     options:
       show_symbol_type_toc: true
