+++ src/centimators/model_estimators/__init__.py
@@ -22,11 +22,11 @@ __all__ = [
 
 _LAZY_IMPORTS: dict[str, str] = {
     # Keras estimators
-    "BaseKerasEstimator": "centimators.model_estimators.keras_estimators",
-    "SequenceEstimator": "centimators.model_estimators.keras_estimators",
-    "MLPRegressor": "centimators.model_estimators.keras_estimators",
-    "BottleneckEncoder": "centimators.model_estimators.keras_estimators",
-    "LSTMRegressor": "centimators.model_estimators.keras_estimators",
+    "BaseKerasEstimator": "centimators.model_estimators.keras_estimators.base",
+    "SequenceEstimator": "centimators.model_estimators.keras_estimators.sequence",
+    "MLPRegressor": "centimators.model_estimators.keras_estimators.dense",
+    "BottleneckEncoder": "centimators.model_estimators.keras_estimators.autoencoder",
+    "LSTMRegressor": "centimators.model_estimators.keras_estimators.sequence",
     # DSPy estimator
     "DSPyMator": "centimators.model_estimators.dspymator",
     # Meta-estimator
